requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Environment Variables
python-dotenv==1.0.0

# Scheduling
APScheduler==3.10.4

# Database
sqlalchemy==2.0.23

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
